Entropy-l-diversity: Add Grassberger entropy estimator
It is now possible to choose the entropy estimator
during the instantiation of an EntropyLDiversity
criterion.

The default is, as was the only option before, the
naive Shannon estimator. This also reflects the original
definition of entropy-l-diversity.

The newly added option is to use the Grassberger estimator, see
S Stammler, S Katzenbeisser, K Hamacher.
Correcting Finite Sampling Issues in Entropy l-diversity.
PSD 2016. LNCS Vol. 9867