Create prototype view and API for data masking
This implements a new masking view, which can be used to configure data
masking. For now it is very rudimentary, only outlining the envisioned
concept. It needs further refinements to be actually usable.

Along with the GUI some basic elements have been defined on the API
level. It is not fully usable yet, but there are some dependecies
between the GUI and API, so they are developed hand in hand.