Fix calculation of row height for data tables for real
This is a first attempt to fix issues with the row height in data tables (#80).
It changes the ColumnHeaderConfiguration as was suggested by upstream [1]. This
is not final, though, since there are now problems with the background image as
well as the image within the cell itself (yellow dot over the label).

[1]: https://bugs.eclipse.org/bugs/show_bug.cgi?id=511784#c8