Remove variable views and generalize support for masking types
This removes the variable views, since they will be replaced by a more
generic masking view. Perturbation will be a particular type of masking,
rather than being treated differently. This patch also adds basic API
support for masking types and changes the prototype GUI so it is compatible
with the API.