Re-introduce concept of random variables and re-arrange layout
Parameter configuration will be implemented in Composites. A map will define
which composite should be displayed for which masking type. Masking parameters
will simply be added to the MaskingConfiguration, since there can also be
optional parameters, etc.

This commit also re-introduces the concept of random variables, which can be
added and configured by the user and assigned to a particular masking. The
views have been prototyped and can exchange data with each other, now they
to be implemented for real.