More prototype functionality in masking perspective
This adds more prototype functionality to the masking perspective, i.e.
an example chart, etc. It is meant to demonstrate the general concept,
but is not yet usable. It also gets rid of the configuration views for
masking operations and random variables, since this will be done using
dialogs and wizards.