Implement new prototype of variable dialog
This refactors some elements of the variable configuration and implements a
first prototype of the variable dialog. It can be used to add and configure
random variables, i.e. select the distribution and parameters. This approach
is more intuitive for the user and can be handled in a better way code-wide.