Generalize distribution parameters and remove specific composites from GUI
This generalizes the parameter concept for a distribution in such a way
that  there is no need for a specific composite for each distribution.
Rather distributions are specified as data types, which have parameters
that can be attached to a random variable. In this way it is rather easy
to display the data, as well as create an appropriate distribution.