Add colored icons for attributeTypes in view 'Classification accuracy'.
In the view 'Analyze/enhance utility' -> 'Classification accuracy' I
added a column to both tables 'Features' and 'Classes' with colored
icons representing the attribute type (correlating to the colors used in
the input/output as well as attribute metadata tables). The icons are
being updated when the attribute type of an attribute is being changed
by the user.