Validate file import for "attribute value order selection dialog"
It is validated if the selected file contains the values of the
attribute's domain. If the file contains no or too few valid values a
warning dialog is displayed. If the file contains more values these
values are ignored and an info dialog is shown to the user.
Therefore, a warning dialog from JFace has been added to the controller.