Add tab 'Classification accuracy' in output side (analyzing view)
Issue: Currently, the feature and class selection for the classification
analysis in the 'Analyze/enhance utility' view is handled via the tab
'Classification accuracy' on the left-hand 'input side', and it is used
for both analysis of the input and output datasets.

Enhancement: In order to make it clearer to the user that this selection is being applied to both input and output datasets, I added the tab 'Classification accuracy' to the right-hand 'output side' with a hint message explaining this fact.